% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT data set.}
\format{
A data frame with 144 rows and 5 variables:
\describe{
  \item{LENGTH}{length in cm}
  \item{WEIGHT}{weight of the fish, in gms}
  \item{DDT}{ddt amount}
  \item{MILE}{length in miles}
  \item{RIVER}{type of river}
  \item{SPECIES}{type of species}
}
}
\source{
\url{https://www.crcpress.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
ddt
}
\description{
A data set collected by the US forces from the tennessee river and its tributaries
}
\keyword{datasets}
