% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Sampling function from Lab 5}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{The number of iterations}

\item{n}{The number of trials}

\item{p}{The probability}
}
\value{
A graph of the probabilities for the trials
}
\description{
Sampling function from Lab 5
}
\examples{
\dontrun{mybin(iter=1000,n=18, p=0.3)}
}
