% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2
This function creates a histrogram of the confidence intervals}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations to perform}

\item{x}{vector of elements}

\item{fun}{the function to apply to the sample}

\item{alpha}{the confidence level for the confidence interval}

\item{cx}{the size of the text on histogram}

\item{...}{other parameters to past to histogram if necessary}
}
\value{
histogram of plotted confidence interval
}
\description{
myboot2
This function creates a histrogram of the confidence intervals
}
\examples{
myboot2(x = rnorm(20,mean=10,sd=4), fun = "mean", alpha = 0.05)
}
