% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{Title mycltp}
\usage{
mycltp(n, iter, lambda = 10)
}
\arguments{
\item{n}{size}

\item{iter}{iterations}

\item{lambda}{mean number of events}
}
\value{
generates visualization of data for a histogram of the sample mean with a theoretical normal curve,
a barplot of the relative frequency of sampeld values `y` from the Poisson distribution, and
a probability function plot for the Poisson distribution with the specified `lambda`
}
\description{
Title mycltp
}
\examples{
mycltp(n=2, iter=10000,lambda=4)
}
