% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Maximum Log-Likelihood Estimation

Description:
This function estimates the maximum log-likelihood for a given likelihood function
and a given dataset. It also produces a plot of the log-likelihood function with
respect to the parameter values.

Usage:
mymaxlik(lfun, x, param, ...)

Arguments:}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{A likelihood function that takes two arguments: a data point and a parameter value.
The function should return the log-likelihood for the given data point and parameter value.}

\item{x}{A vector of data points for which the maximum log-likelihood should be estimated.}

\item{param}{A vector of parameter values at which the likelihood function should be evaluated.}

\item{...}{Additional graphical parameters to be passed to the plot function.}
}
\value{
A list containing the following elements:
- i:      The index of the maximum log-likelihood in the 'y' vector.
- parami: The parameter value that corresponds to the maximum log-likelihood.
- yi:     The maximum log-likelihood value.
- slope:  A vector of slopes around the maximum log-likelihood point. This can be used
          as a diagnostic tool to check if the slope changes sign around the maximum point.
          If the slope changes sign from positive to negative, this indicates a maximum.
}
\description{
Maximum Log-Likelihood Estimation

Description:
This function estimates the maximum log-likelihood for a given likelihood function
and a given dataset. It also produces a plot of the log-likelihood function with
respect to the parameter values.

Usage:
mymaxlik(lfun, x, param, ...)

Arguments:
}
\examples{

logbin=function(x,param) log(dbinom(x,prob=param,size=20))
y <- c(3,3,4,3,4,5,5,4)
mymaxlik(lfun = logbin, x = y, param = seq(0, 1, length = 1000), main = "Binomial, n = 20")
}
