% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(a, mu, sigma)
}
\arguments{
\item{a}{Upper bound}

\item{mu}{mean}

\item{sigma}{standard deviation}
}
\value{
plotted curve with area shaded in
}
\description{
myncurve
}
\examples{
myncurve(a=1, mu=2, sigma=3)
}
