% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myntickets.R
\name{ntickets}
\alias{ntickets}
\title{Title Find the minimum amount of tickets needed}
\usage{
ntickets(N = 200, gamma = 0.02, p = 0.95)
}
\arguments{
\item{N}{number of seats}

\item{gamma}{probability the plane will truly be overbooked}

\item{p}{probability of a "show"}
}
\value{
a continuous and discrete graph and a list containing the discrete distribution
and normal approximation
}
\description{
Title Find the minimum amount of tickets needed
}
\examples{
ntickets(400, 0.02, 0.95)
}
