% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunc.R
\name{myfunc}
\alias{myfunc}
\title{myfunc
This function is for a quadratic}
\usage{
myfunc(x)
}
\arguments{
\item{x}{variable as an integer}
}
\value{
it returns a line of the equation
}
\description{
myfunc
This function is for a quadratic
}
\examples{
myfunc(1)

}
