% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myhyper.R
\name{myhyper}
\alias{myhyper}
\title{myhyper
myhyper function from lab 5.
This function forms a barplot of the trials and successes}
\usage{
myhyper(iter = 100, N = 20, r = 12, n = 5)
}
\arguments{
\item{iter}{iterations}

\item{N}{size}

\item{r}{amount of a certain choice}

\item{n}{amount of times you'll choose}
}
\value{
results as a barplot
}
\description{
myhyper
myhyper function from lab 5.
This function forms a barplot of the trials and successes
}
\examples{
myhyper(100, 20, 12, 5)
}
