% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysample.R
\name{mysample}
\alias{mysample}
\title{mysample}
\usage{
mysample(n = 1, iter = 10, time = 0.5)
}
\arguments{
\item{n}{desc}

\item{iter}{desc}

\item{time}{desc}
}
\value{
desc
}
\description{
mysample function from lab 5.
This is my function.
This is what it does.
}
\examples{
mysample(1, 10, 0.5)
}
